+++ src/BShop.API/Data/Models/ProductCategory.cs
@@ -2,11 +2,11 @@
 {
     public class ProductCategory
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public Product Product { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public Category Category { get; set; }
     }
